Add integration into the Provider Code
